chore: xlsx as devDependency and move education seeding to separate file
